refactored data fetching "return a future"
In an effort to separate the final result genereation from the http
request this aims to update the functions to return futures.

My hunch is this is best done in the most general way possible.
Currently I'm returning a future of Vec<T> where T is a struct that
serde knows how to deserialize. If I generalize further and simply
return a future of serde_json::Value, then call and_then() on it to
parse the body... I think we'd be able to simplify the codebase overall.

Additionally, I think the uri building needs to be formalized a bit more
than it is right now. Maybe factorying the uri building parts into their
own struct, then attach it to the MarvelClient... I'll try it and see.